Adds IODataModelAdapter to refactor away model creation
This adds a layer of indirection to the reduce the model parsing calls to only
one per cached URI. The IODataModelAdapter follows the IODataAdapter pattern
and is used to create the model from the parsed metadata. The IODataAdapter then
marries the ISession and IODataModelAdapter to create the ODataAdapter for the
given version of the OData protocol.